If the backup server known to a PBServer changes to "", no attempt will be made to transfer key values to that (nonexistant) server any more. Also, now only the Primary server can call transfer_kvstore (earlier iterations erroneously allowed the primary and backup to transfer to each other which caused a race condition to see which copy of kvstore was actually persisted across both). Also, no locks are needed inside transfer_kvstore since it is enclosed in the PBServer lock taken out by the tick method.
Attempts to transter kvstore data continue until they succeed to handle network errors and the case where the backup server has not talked to the viewservice yet and still thinks it is in an idle role. The Receive_kvstore method of PBServer now only accepts kvstore if the server is in a backup role. If it is in any other role, it replies with ErrWrongServer and the lock during that request is released - it can then eventually communicate with the viewservice and an idle server that has been promoted will learn that it is now a backup and accept the repeatedly incoming transfer_kvstore RPC. This also prevents a random idle server or the primary from accepting kvstore data contents

Note that Receive_kvstore will always accept and replace its kvstore contents if the server is running in a backup role, meaning malicious external clients could erase data on the backup and the Primary/Backup synchronization would be violated.