Wrote PBServer methods called InternalGet and InternalPut for PBServers to send RPC requests to one another to allow the Primary server to forward requests to the Backup and wait for a response. Whenever a Primary receives a Get/Put request from a client and there is a backup present, the request is forwarded to the backup and the primary waits for the response before making any state changes. If the response is not ErrWrongServer then the request was successful and the Primary may proceed with processing the response. Otherwise an error reply is sent back to the client indicating ErrWrongServer (which may mean only that the Backup is wrong, not that the Primary is wrong).
Notice that the Primary does not retry to forward to the Backup as the Primary needs processor time in requests in order for tick to be called and for the Primary to realize it is no longer primary, the backup has changed, etc. It is the client's responsibility to retry requests (which should only go to the Primary).